requirements.txt
ctrader-open-api>=0.9.0,<1.0
twisted>=24.0,<25.0
yfinance>=0.2.40,<0.3
pandas>=2.2,<3.0
numpy>=1.26,<2.0
scikit-learn>=1.5,<2.0
xgboost>=2.1,<3.0
lightgbm>=4.5,<5.0
shap>=0.46,<1.0
smartmoneyconcepts>=0.0.22,<1.0
pandas-ta>=0.3.14b1,<1.0
python-telegram-bot>=20.0,<22.0
groq>=0.11,<1.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
schedule>=1.2,<2.0
sqlalchemy>=2.0,<3.0
pyarrow>=17.0,<19.0
fastparquet>=2024.5,<2026.0
matplotlib>=3.9,<4.0
seaborn>=0.13,<1.0
jupyter>=1.1,<2.0
notebook>=7.2,<8.0
